BfHaskell.StreamingAPI.Model

Description

Models

DateTime Formatting

newtype DateTime

Constructors

DateTime

Fields

unDateTime :: UTCTime

_readDateTime :: (ParseTime t, Monad m, Alternative m) => String -> m t

_parseISO8601

_showDateTime :: (t ~ UTCTime, FormatTime t) => t -> String

TI.formatISO8601Millis

_parseISO8601 :: (ParseTime t, Monad m, Alternative m) => String -> m t

parse an ISO8601 date-time string

_omitNulls :: [(Text, Value)] -> Value

AllRequestTypesExample

data AllRequestTypesExample

AllRequestTypesExample

Constructors

AllRequestTypesExample

Fields

allRequestTypesExampleOpTypes :: !(Maybe E'OpTypes)

"opTypes"

allRequestTypesExampleHeartbeat :: !(Maybe HeartbeatMessage)

"heartbeat"

allRequestTypesExampleOrderSubscriptionMessage :: !(Maybe OrderSubscriptionMessage)

"orderSubscriptionMessage"

allRequestTypesExampleMarketSubscription :: !(Maybe MarketSubscriptionMessage)

"marketSubscription"

allRequestTypesExampleAuthentication :: !(Maybe AuthenticationMessage)

"authentication"

mkAllRequestTypesExample :: AllRequestTypesExample

Construct a value of type AllRequestTypesExample (by applying it's required fields, if any)

data StreamResponse

StreamResponse

Constructors

SRMarketChangeMessage MarketChangeMessage

SRConnectionMessage ConnectionMessage

SROrderChangeMessage OrderChangeMessage

SRStatusMessage StatusMessage

AllResponseTypesExample

data AllResponseTypesExample

AllResponseTypesExample

Constructors

AllResponseTypesExample

Fields

allResponseTypesExampleOpTypes :: !(Maybe E'OpTypes2)

"opTypes"

allResponseTypesExampleMarketChangeMessage :: !(Maybe MarketChangeMessage)

"marketChangeMessage"

allResponseTypesExampleConnection :: !(Maybe ConnectionMessage)

"connection"

allResponseTypesExampleOrderChangeMessage :: !(Maybe OrderChangeMessage)

"orderChangeMessage"

allResponseTypesExampleStatus :: !(Maybe StatusMessage)

"status"

mkAllResponseTypesExample :: AllResponseTypesExample

Construct a value of type AllResponseTypesExample (by applying it's required fields, if any)

KeyLineDefinition

data KeyLineDefinition

KeyLineDefinition

Constructors

KeyLineDefinition

Fields

keyLineDefinitionKl :: !(Maybe [KeyLineSelection])

"kl"

mkKeyLineDefinition :: KeyLineDefinition

Construct a value of type KeyLineDefinition (by applying it's required fields, if any)

KeyLineSelection

data KeyLineSelection

KeyLineSelection

Constructors

KeyLineSelection

Fields

keyLineSelectionId :: !(Maybe Integer)

"id"

keyLineSelectionHc :: !(Maybe Double)

"hc"

mkKeyLineSelection :: KeyLineSelection

Construct a value of type KeyLineSelection (by applying it's required fields, if any)

MarketChange

data MarketChange

MarketChange

Constructors

MarketChange

Fields

marketChangeRc :: !(Maybe [RunnerChange])

"rc" - Runner Changes - a list of changes to runners (or null if un-changed)

marketChangeImg :: !(Maybe Bool)

"img" - Image - replace existing prices / data with the data supplied: it is not a delta (or null if delta)

marketChangeTv :: !(Maybe Double)

"tv" - The total amount matched across the market. This value is truncated at 2dp (or null if un-changed)

marketChangeCon :: !(Maybe Bool)

"con" - Conflated - have more than a single change been combined (or null if not conflated)

marketChangeMarketDefinition :: !(Maybe MarketDefinition)

"marketDefinition" - Market Definition - the definition of the market (or null if un-changed)

marketChangeId :: !(Maybe Text)

"id" - Market Id - the id of the market

mkMarketChange :: MarketChange

Construct a value of type MarketChange (by applying it's required fields, if any)

MarketDataFilter

data MarketDataFilter

MarketDataFilter

Constructors

MarketDataFilter

Fields

marketDataFilterLadderLevels :: !(Maybe Int)

"ladderLevels"

marketDataFilterFields :: !(Maybe [E'Fields])

"fields"

mkMarketDataFilter :: MarketDataFilter

Construct a value of type MarketDataFilter (by applying it's required fields, if any)

MarketDefinition

data MarketDefinition

MarketDefinition

Constructors

MarketDefinition

Fields

marketDefinitionVenue :: !(Maybe Text)

"venue"

marketDefinitionRaceType :: !(Maybe Text)

"raceType"

marketDefinitionSettledTime :: !(Maybe DateTime)

"settledTime"

marketDefinitionTimezone :: !(Maybe Text)

"timezone"

marketDefinitionEachWayDivisor :: !(Maybe Double)

"eachWayDivisor"

marketDefinitionRegulators :: !(Maybe [Text])

"regulators" - The market regulators.

marketDefinitionMarketType :: !(Maybe Text)

"marketType"

marketDefinitionMarketBaseRate :: !(Maybe Double)

"marketBaseRate"

marketDefinitionNumberOfWinners :: !(Maybe Int)

"numberOfWinners"

marketDefinitionCountryCode :: !(Maybe Text)

"countryCode"

marketDefinitionLineMaxUnit :: !(Maybe Double)

"lineMaxUnit" - For Handicap and Line markets, the maximum value for the outcome, in market units for this market (eg 100 runs).

marketDefinitionInPlay :: !(Maybe Bool)

"inPlay"

marketDefinitionBetDelay :: !(Maybe Int)

"betDelay"

marketDefinitionBspMarket :: !(Maybe Bool)

"bspMarket"

marketDefinitionBettingType :: !(Maybe E'BettingTypes)

"bettingType"

marketDefinitionNumberOfActiveRunners :: !(Maybe Int)

"numberOfActiveRunners"

marketDefinitionLineMinUnit :: !(Maybe Double)

"lineMinUnit" - For Handicap and Line markets, the minimum value for the outcome, in market units for this market (eg 0 runs).

marketDefinitionEventId :: !(Maybe Text)

"eventId"

marketDefinitionCrossMatching :: !(Maybe Bool)

"crossMatching"

marketDefinitionRunnersVoidable :: !(Maybe Bool)

"runnersVoidable"

marketDefinitionTurnInPlayEnabled :: !(Maybe Bool)

"turnInPlayEnabled"

marketDefinitionPriceLadderDefinition :: !(Maybe PriceLadderDefinition)

"priceLadderDefinition" - Definition of the price ladder type and any related data.

marketDefinitionKeyLineDefinition :: !(Maybe KeyLineDefinition)

"keyLineDefinition" - Definition of a markets key line selection (for valid markets), comprising the selectionId and handicap of the team it is applied to.

marketDefinitionSuspendTime :: !(Maybe DateTime)

"suspendTime"

marketDefinitionDiscountAllowed :: !(Maybe Bool)

"discountAllowed"

marketDefinitionPersistenceEnabled :: !(Maybe Bool)

"persistenceEnabled"

marketDefinitionRunners :: !(Maybe [RunnerDefinition])

"runners"

marketDefinitionVersion :: !(Maybe Integer)

"version"

marketDefinitionEventTypeId :: !(Maybe Text)

"eventTypeId" - The Event Type the market is contained within.

marketDefinitionComplete :: !(Maybe Bool)

"complete"

marketDefinitionOpenDate :: !(Maybe DateTime)

"openDate"

marketDefinitionMarketTime :: !(Maybe DateTime)

"marketTime"

marketDefinitionBspReconciled :: !(Maybe Bool)

"bspReconciled"

marketDefinitionLineInterval :: !(Maybe Double)

"lineInterval" - For Handicap and Line markets, the lines available on this market will be between the range of lineMinUnit and lineMaxUnit, in increments of the lineInterval value. e.g. If unit is runs, lineMinUnit=10, lineMaxUnit=20 and lineInterval=0.5, then valid lines include 10, 10.5, 11, 11.5 up to 20 runs.

marketDefinitionStatus :: !(Maybe E'Status)

"status"

mkMarketDefinition :: MarketDefinition

Construct a value of type MarketDefinition (by applying it's required fields, if any)

MarketFilter

data MarketFilter

MarketFilter

Constructors

MarketFilter

Fields

marketFilterCountryCodes :: !(Maybe [Text])

"countryCodes"

marketFilterBettingTypes :: !(Maybe [E'BettingTypes])

"bettingTypes"

marketFilterTurnInPlayEnabled :: !(Maybe Bool)

"turnInPlayEnabled"

marketFilterMarketTypes :: !(Maybe [Text])

"marketTypes"

marketFilterVenues :: !(Maybe [Text])

"venues"

marketFilterMarketIds :: !(Maybe [Text])

"marketIds"

marketFilterEventTypeIds :: !(Maybe [Text])

"eventTypeIds"

marketFilterEventIds :: !(Maybe [Text])

"eventIds"

marketFilterBspMarket :: !(Maybe Bool)

"bspMarket"

marketFilterRaceTypes :: !(Maybe [Text])

"raceTypes"

mkMarketFilter :: MarketFilter

Construct a value of type MarketFilter (by applying it's required fields, if any)

Order

data Order

Order

Constructors

Order

Fields

orderSide :: !(Maybe E'Side)

"side" - Side - the side of the order. For Line markets a 'B' bet refers to a SELL line and an 'L' bet refers to a BUY line.

orderSv :: !(Maybe Double)

"sv" - Size Voided - the amount of the order that has been voided

orderPt :: !(Maybe E'Pt)

"pt" - Persistence Type - whether the order will persist at in play or not (L = LAPSE, P = PERSIST, MOC = Market On Close)

orderOt :: !(Maybe E'Ot)

"ot" - Order Type - the type of the order (L = LIMIT, MOC = MARKET_ON_CLOSE, LOC = LIMIT_ON_CLOSE)

orderP :: !(Maybe Double)

"p" - Price - the original placed price of the order. Line markets operate at even-money odds of 2.0. However, price for these markets refers to the line positions available as defined by the markets min-max range and interval steps

orderSc :: !(Maybe Double)

"sc" - Size Cancelled - the amount of the order that has been cancelled

orderRc :: !(Maybe Text)

"rc" - Regulator Code - the regulator of the order

orderS :: !(Maybe Double)

"s" - Size - the original placed size of the order

orderPd :: !(Maybe Integer)

"pd" - Placed Date - the date the order was placed

orderRac :: !(Maybe Text)

"rac" - Regulator Auth Code - the auth code returned by the regulator

orderMd :: !(Maybe Integer)

"md" - Matched Date - the date the order was matched (null if the order is not matched)

orderLd :: !(Maybe Integer)

"ld" - Lapsed Date - the date the order was lapsed (null if the order is not lapsed)

orderSl :: !(Maybe Double)

"sl" - Size Lapsed - the amount of the order that has been lapsed

orderAvp :: !(Maybe Double)

"avp" - Average Price Matched - the average price the order was matched at (null if the order is not matched). This value is not meaningful for activity on Line markets and is not guaranteed to be returned or maintained for these markets.

orderSm :: !(Maybe Double)

"sm" - Size Matched - the amount of the order that has been matched

orderRfo :: !(Maybe Text)

"rfo" - Order Reference - the customer's order reference for this order (empty string if one was not set)

orderId :: !(Maybe Text)

"id" - Bet Id - the id of the order

orderBsp :: !(Maybe Double)

"bsp" - BSP Liability - the BSP liability of the order (null if the order is not a BSP order)

orderRfs :: !(Maybe Text)

"rfs" - Strategy Reference - the customer's strategy reference for this order (empty string if one was not set)

orderStatus :: !(Maybe E'Status3)

"status" - Status - the status of the order (E = EXECUTABLE, EC = EXECUTION_COMPLETE)

orderSr :: !(Maybe Double)

"sr" - Size Remaining - the amount of the order that is remaining unmatched

mkOrder :: Order

Construct a value of type Order (by applying it's required fields, if any)

OrderFilter

data OrderFilter

OrderFilter

Constructors

OrderFilter

Fields

orderFilterIncludeOverallPosition :: !(Maybe Bool)

"includeOverallPosition" - Returns overall net position (See: OrderRunnerChange.mb OrderRunnerChange.ml). Default=true

orderFilterAccountIds :: !(Maybe [Integer])

"accountIds" - Internal use only &amp; should not be set on your filter (your subscription is already locked to your account). If set subscription will fail.

orderFilterCustomerStrategyRefs :: !(Maybe [Text])

"customerStrategyRefs" - Restricts to specified customerStrategyRefs; this will filter orders and StrategyMatchChanges accordingly (Note: overall postition is not filtered)

orderFilterPartitionMatchedByStrategyRef :: !(Maybe Bool)

"partitionMatchedByStrategyRef" - Returns strategy positions (See: OrderRunnerChange.smc=Map&lt;customerStrategyRef, StrategyMatchChange&gt;) - these are sent in delta format as per overall position. Default=false

mkOrderFilter :: OrderFilter

Construct a value of type OrderFilter (by applying it's required fields, if any)

OrderMarketChange

data OrderMarketChange

OrderMarketChange

Constructors

OrderMarketChange

Fields

orderMarketChangeAccountId :: !(Maybe Integer)

"accountId"

orderMarketChangeOrc :: !(Maybe [OrderRunnerChange])

"orc" - Order Changes - a list of changes to orders on a selection

orderMarketChangeClosed :: !(Maybe Bool)

"closed"

orderMarketChangeId :: !(Maybe Text)

"id" - Market Id - the id of the market the order is on

mkOrderMarketChange :: OrderMarketChange

Construct a value of type OrderMarketChange (by applying it's required fields, if any)

OrderRunnerChange

data OrderRunnerChange

OrderRunnerChange

Constructors

OrderRunnerChange

Fields

orderRunnerChangeMb :: !(Maybe [[Double]])

"mb" - Matched Backs - matched amounts by distinct matched price on the Back side for this runner (selection)

orderRunnerChangeSmc :: !(Maybe (Map String StrategyMatchChange))

"smc" - Strategy Matches - Matched Backs and Matched Lays grouped by strategy reference

orderRunnerChangeUo :: !(Maybe [Order])

"uo" - Unmatched Orders - orders on this runner (selection) that are not fully matched

orderRunnerChangeId :: !(Maybe Integer)

"id" - Selection Id - the id of the runner (selection)

orderRunnerChangeHc :: !(Maybe Double)

"hc" - Handicap - the handicap of the runner (selection) (null if not applicable)

orderRunnerChangeFullImage :: !(Maybe Bool)

"fullImage"

orderRunnerChangeMl :: !(Maybe [[Double]])

"ml" - Matched Lays - matched amounts by distinct matched price on the Lay side for this runner (selection)

mkOrderRunnerChange :: OrderRunnerChange

Construct a value of type OrderRunnerChange (by applying it's required fields, if any)

PriceLadderDefinition

data PriceLadderDefinition

PriceLadderDefinition

Constructors

PriceLadderDefinition

Fields

priceLadderDefinitionType :: !(Maybe E'Type)

"type"

mkPriceLadderDefinition :: PriceLadderDefinition

Construct a value of type PriceLadderDefinition (by applying it's required fields, if any)

RequestMessage

data RequestMessage

RequestMessage

Constructors

RequestMessage

Fields

requestMessageOp :: !(Maybe Text)

"op" - The operation type

requestMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

mkRequestMessage :: RequestMessage

Construct a value of type RequestMessage (by applying it's required fields, if any)

ResponseMessage

data ResponseMessage

ResponseMessage

Constructors

ResponseMessage

Fields

responseMessageOp :: !(Maybe Text)

"op" - The operation type

responseMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

mkResponseMessage :: ResponseMessage

Construct a value of type ResponseMessage (by applying it's required fields, if any)

RunnerChange

data RunnerChange

RunnerChange

Constructors

RunnerChange

Fields

runnerChangeTv :: !(Maybe Double)

"tv" - The total amount matched. This value is truncated at 2dp.

runnerChangeBatb :: !(Maybe [[Double]])

"batb" - Best Available To Back - LevelPriceVol triple delta of price changes, keyed by level (0 vol is remove)

runnerChangeSpb :: !(Maybe [[Double]])

"spb" - Starting Price Back - PriceVol tuple delta of price changes (0 vol is remove)

runnerChangeBdatl :: !(Maybe [[Double]])

"bdatl" - Best Display Available To Lay (includes virtual prices)- LevelPriceVol triple delta of price changes, keyed by level (0 vol is remove)

runnerChangeTrd :: !(Maybe [[Double]])

"trd" - Traded - PriceVol tuple delta of price changes (0 vol is remove)

runnerChangeSpf :: !(Maybe Double)

"spf" - Starting Price Far - The far starting price (or null if un-changed)

runnerChangeLtp :: !(Maybe Double)

"ltp" - Last Traded Price - The last traded price (or null if un-changed)

runnerChangeAtb :: !(Maybe [[Double]])

"atb" - Available To Back - PriceVol tuple delta of price changes (0 vol is remove)

runnerChangeSpl :: !(Maybe [[Double]])

"spl" - Starting Price Lay - PriceVol tuple delta of price changes (0 vol is remove)

runnerChangeSpn :: !(Maybe Double)

"spn" - Starting Price Near - The far starting price (or null if un-changed)

runnerChangeAtl :: !(Maybe [[Double]])

"atl" - Available To Lay - PriceVol tuple delta of price changes (0 vol is remove)

runnerChangeBatl :: !(Maybe [[Double]])

"batl" - Best Available To Lay - LevelPriceVol triple delta of price changes, keyed by level (0 vol is remove)

runnerChangeId :: !(Maybe Integer)

"id" - Selection Id - the id of the runner (selection)

runnerChangeHc :: !(Maybe Double)

"hc" - Handicap - the handicap of the runner (selection) (null if not applicable)

runnerChangeBdatb :: !(Maybe [[Double]])

"bdatb" - Best Display Available To Back (includes virtual prices)- LevelPriceVol triple delta of price changes, keyed by level (0 vol is remove)

mkRunnerChange :: RunnerChange

Construct a value of type RunnerChange (by applying it's required fields, if any)

RunnerDefinition

data RunnerDefinition

RunnerDefinition

Constructors

RunnerDefinition

Fields

runnerDefinitionSortPriority :: !(Maybe Int)

"sortPriority"

runnerDefinitionRemovalDate :: !(Maybe DateTime)

"removalDate"

runnerDefinitionId :: !(Maybe Integer)

"id" - Selection Id - the id of the runner (selection)

runnerDefinitionHc :: !(Maybe Double)

"hc" - Handicap - the handicap of the runner (selection) (null if not applicable)

runnerDefinitionAdjustmentFactor :: !(Maybe Double)

"adjustmentFactor"

runnerDefinitionBsp :: !(Maybe Double)

"bsp"

runnerDefinitionStatus :: !(Maybe E'Status2)

"status"

mkRunnerDefinition :: RunnerDefinition

Construct a value of type RunnerDefinition (by applying it's required fields, if any)

StrategyMatchChange

data StrategyMatchChange

StrategyMatchChange

Constructors

StrategyMatchChange

Fields

strategyMatchChangeMb :: !(Maybe [[Double]])

"mb" - Matched Backs - matched amounts by distinct matched price on the Back side for this strategy

strategyMatchChangeMl :: !(Maybe [[Double]])

"ml" - Matched Lays - matched amounts by distinct matched price on the Lay side for this strategy

mkStrategyMatchChange :: StrategyMatchChange

Construct a value of type StrategyMatchChange (by applying it's required fields, if any)

AuthenticationMessage

data AuthenticationMessage

AuthenticationMessage

Constructors

AuthenticationMessage

Fields

authenticationMessageOp :: !(Maybe Text)

"op" - The operation type

authenticationMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

authenticationMessageSession :: !(Maybe Text)

"session"

authenticationMessageAppKey :: !(Maybe Text)

"appKey"

mkAuthenticationMessage :: AuthenticationMessage

Construct a value of type AuthenticationMessage (by applying it's required fields, if any)

ConnectionMessage

data ConnectionMessage

ConnectionMessage

Constructors

ConnectionMessage

Fields

connectionMessageOp :: !(Maybe Text)

"op" - The operation type

connectionMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

connectionMessageConnectionId :: !(Maybe Text)

"connectionId" - The connection id

mkConnectionMessage :: ConnectionMessage

Construct a value of type ConnectionMessage (by applying it's required fields, if any)

HeartbeatMessage

data HeartbeatMessage

HeartbeatMessage

Constructors

HeartbeatMessage

Fields

heartbeatMessageOp :: !(Maybe Text)

"op" - The operation type

heartbeatMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

mkHeartbeatMessage :: HeartbeatMessage

Construct a value of type HeartbeatMessage (by applying it's required fields, if any)

MarketChangeMessage

data MarketChangeMessage

MarketChangeMessage

Constructors

MarketChangeMessage

Fields

marketChangeMessageOp :: !(Maybe Text)

"op" - The operation type

marketChangeMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

marketChangeMessageCt :: !(Maybe E'Ct)

"ct" - Change Type - set to indicate the type of change - if null this is a delta)

marketChangeMessageClk :: !(Maybe Text)

"clk" - Token value (non-null) should be stored and passed in a MarketSubscriptionMessage to resume subscription (in case of disconnect)

marketChangeMessageHeartbeatMs :: !(Maybe Integer)

"heartbeatMs" - Heartbeat Milliseconds - the heartbeat rate (may differ from requested: bounds are 500 to 30000)

marketChangeMessagePt :: !(Maybe Integer)

"pt" - Publish Time (in millis since epoch) that the changes were generated

marketChangeMessageInitialClk :: !(Maybe Text)

"initialClk" - Token value (non-null) should be stored and passed in a MarketSubscriptionMessage to resume subscription (in case of disconnect)

marketChangeMessageMc :: !(Maybe [MarketChange])

"mc" - MarketChanges - the modifications to markets (will be null on a heartbeat

marketChangeMessageConflateMs :: !(Maybe Integer)

"conflateMs" - Conflate Milliseconds - the conflation rate (may differ from that requested if subscription is delayed)

marketChangeMessageSegmentType :: !(Maybe E'SegmentType)

"segmentType" - Segment Type - if the change is split into multiple segments, this denotes the beginning and end of a change, and segments in between. Will be null if data is not segmented

marketChangeMessageStatus :: !(Maybe Int)

"status" - Stream status: set to null if the exchange stream data is up to date and 503 if the downstream services are experiencing latencies

mkMarketChangeMessage :: MarketChangeMessage

Construct a value of type MarketChangeMessage (by applying it's required fields, if any)

MarketSubscriptionMessage

data MarketSubscriptionMessage

MarketSubscriptionMessage

Constructors

MarketSubscriptionMessage

Fields

marketSubscriptionMessageOp :: !(Maybe Text)

"op" - The operation type

marketSubscriptionMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

marketSubscriptionMessageSegmentationEnabled :: !(Maybe Bool)

"segmentationEnabled" - Segmentation Enabled - allow the server to send large sets of data in segments, instead of a single block

marketSubscriptionMessageClk :: !(Maybe Text)

"clk" - Token value delta (received in MarketChangeMessage) that should be passed to resume a subscription

marketSubscriptionMessageHeartbeatMs :: !(Maybe Integer)

"heartbeatMs" - Heartbeat Milliseconds - the heartbeat rate (looped back on initial image after validation: bounds are 500 to 5000)

marketSubscriptionMessageInitialClk :: !(Maybe Text)

"initialClk" - Token value (received in initial MarketChangeMessage) that should be passed to resume a subscription

marketSubscriptionMessageMarketFilter :: !(Maybe MarketFilter)

"marketFilter"

marketSubscriptionMessageConflateMs :: !(Maybe Integer)

"conflateMs" - Conflate Milliseconds - the conflation rate (looped back on initial image after validation: bounds are 0 to 120000)

marketSubscriptionMessageMarketDataFilter :: !(Maybe MarketDataFilter)

"marketDataFilter"

mkMarketSubscriptionMessage :: MarketSubscriptionMessage

Construct a value of type MarketSubscriptionMessage (by applying it's required fields, if any)

OrderChangeMessage

data OrderChangeMessage

OrderChangeMessage

Constructors

OrderChangeMessage

Fields

orderChangeMessageOp :: !(Maybe Text)

"op" - The operation type

orderChangeMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

orderChangeMessageCt :: !(Maybe E'Ct)

"ct" - Change Type - set to indicate the type of change - if null this is a delta)

orderChangeMessageClk :: !(Maybe Text)

"clk" - Token value (non-null) should be stored and passed in a MarketSubscriptionMessage to resume subscription (in case of disconnect)

orderChangeMessageHeartbeatMs :: !(Maybe Integer)

"heartbeatMs" - Heartbeat Milliseconds - the heartbeat rate (may differ from requested: bounds are 500 to 30000)

orderChangeMessagePt :: !(Maybe Integer)

"pt" - Publish Time (in millis since epoch) that the changes were generated

orderChangeMessageOc :: !(Maybe [OrderMarketChange])

"oc" - OrderMarketChanges - the modifications to account's orders (will be null on a heartbeat

orderChangeMessageInitialClk :: !(Maybe Text)

"initialClk" - Token value (non-null) should be stored and passed in a MarketSubscriptionMessage to resume subscription (in case of disconnect)

orderChangeMessageConflateMs :: !(Maybe Integer)

"conflateMs" - Conflate Milliseconds - the conflation rate (may differ from that requested if subscription is delayed)

orderChangeMessageSegmentType :: !(Maybe E'SegmentType)

"segmentType" - Segment Type - if the change is split into multiple segments, this denotes the beginning and end of a change, and segments in between. Will be null if data is not segmented

orderChangeMessageStatus :: !(Maybe Int)

"status" - Stream status: set to null if the exchange stream data is up to date and 503 if the downstream services are experiencing latencies

mkOrderChangeMessage :: OrderChangeMessage

Construct a value of type OrderChangeMessage (by applying it's required fields, if any)

OrderSubscriptionMessage

data OrderSubscriptionMessage

OrderSubscriptionMessage

Constructors

OrderSubscriptionMessage

Fields

orderSubscriptionMessageOp :: !(Maybe Text)

"op" - The operation type

orderSubscriptionMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

orderSubscriptionMessageSegmentationEnabled :: !(Maybe Bool)

"segmentationEnabled" - Segmentation Enabled - allow the server to send large sets of data in segments, instead of a single block

orderSubscriptionMessageOrderFilter :: !(Maybe OrderFilter)

"orderFilter" - Optional filter applied to order subscription

orderSubscriptionMessageClk :: !(Maybe Text)

"clk" - Token value delta (received in MarketChangeMessage) that should be passed to resume a subscription

orderSubscriptionMessageHeartbeatMs :: !(Maybe Integer)

"heartbeatMs" - Heartbeat Milliseconds - the heartbeat rate (looped back on initial image after validation: bounds are 500 to 5000)

orderSubscriptionMessageInitialClk :: !(Maybe Text)

"initialClk" - Token value (received in initial MarketChangeMessage) that should be passed to resume a subscription

orderSubscriptionMessageConflateMs :: !(Maybe Integer)

"conflateMs" - Conflate Milliseconds - the conflation rate (looped back on initial image after validation: bounds are 0 to 120000)

mkOrderSubscriptionMessage :: OrderSubscriptionMessage

Construct a value of type OrderSubscriptionMessage (by applying it's required fields, if any)

StatusMessage

data StatusMessage

StatusMessage

Constructors

StatusMessage

Fields

statusMessageOp :: !(Maybe Text)

"op" - The operation type

statusMessageId :: !(Maybe Int)

"id" - Client generated unique id to link request with response (like json rpc)

statusMessageErrorMessage :: !(Maybe Text)

"errorMessage" - Additional message in case of a failure

statusMessageErrorCode :: !(Maybe E'ErrorCode)

"errorCode" - The type of error in case of a failure

statusMessageConnectionId :: !(Maybe Text)

"connectionId" - The connection id

statusMessageConnectionClosed :: !(Maybe Bool)

"connectionClosed" - Is the connection now closed

statusMessageStatusCode :: !(Maybe E'StatusCode)

"statusCode" - The status of the last request

mkStatusMessage :: StatusMessage

Construct a value of type StatusMessage (by applying it's required fields, if any)

Enums

E'BettingTypes

data E'BettingTypes

Enum of Text

Constructors

E'BettingTypes'ODDS
ODDS

E'BettingTypes'LINE
LINE

E'BettingTypes'RANGE
RANGE

E'BettingTypes'ASIAN_HANDICAP_DOUBLE_LINE
ASIAN_HANDICAP_DOUBLE_LINE

E'BettingTypes'ASIAN_HANDICAP_SINGLE_LINE
ASIAN_HANDICAP_SINGLE_LINE

fromE'BettingTypes :: E'BettingTypes -> Text

unwrap E'BettingTypes enum

toE'BettingTypes :: Text -> Either String E'BettingTypes

parse E'BettingTypes enum

E'Ct

data E'Ct

Enum of Text . Change Type - set to indicate the type of change - if null this is a delta)

Constructors

E'Ct'SUB_IMAGE
SUB_IMAGE

E'Ct'RESUB_DELTA
RESUB_DELTA

E'Ct'HEARTBEAT
HEARTBEAT

fromE'Ct :: E'Ct -> Text

unwrap E'Ct enum

toE'Ct :: Text -> Either String E'Ct

parse E'Ct enum

E'ErrorCode

data E'ErrorCode

Enum of Text . The type of error in case of a failure

Constructors

E'ErrorCode'NO_APP_KEY
NO_APP_KEY

E'ErrorCode'INVALID_APP_KEY
INVALID_APP_KEY

E'ErrorCode'NO_SESSION
NO_SESSION

E'ErrorCode'INVALID_SESSION_INFORMATION
INVALID_SESSION_INFORMATION

E'ErrorCode'NOT_AUTHORIZED
NOT_AUTHORIZED

E'ErrorCode'INVALID_INPUT
INVALID_INPUT

E'ErrorCode'INVALID_CLOCK
INVALID_CLOCK

E'ErrorCode'UNEXPECTED_ERROR
UNEXPECTED_ERROR

E'ErrorCode'TIMEOUT
TIMEOUT

E'ErrorCode'SUBSCRIPTION_LIMIT_EXCEEDED
SUBSCRIPTION_LIMIT_EXCEEDED

E'ErrorCode'INVALID_REQUEST
INVALID_REQUEST

E'ErrorCode'CONNECTION_FAILED
CONNECTION_FAILED

E'ErrorCode'MAX_CONNECTION_LIMIT_EXCEEDED
MAX_CONNECTION_LIMIT_EXCEEDED

fromE'ErrorCode :: E'ErrorCode -> Text

unwrap E'ErrorCode enum

toE'ErrorCode :: Text -> Either String E'ErrorCode

parse E'ErrorCode enum

E'Fields

data E'Fields

Enum of Text

Constructors

E'Fields'EX_BEST_OFFERS_DISP
EX_BEST_OFFERS_DISP

E'Fields'EX_BEST_OFFERS
EX_BEST_OFFERS

E'Fields'EX_ALL_OFFERS
EX_ALL_OFFERS

E'Fields'EX_TRADED
EX_TRADED

E'Fields'EX_TRADED_VOL
EX_TRADED_VOL

E'Fields'EX_LTP
EX_LTP

E'Fields'EX_MARKET_DEF
EX_MARKET_DEF

E'Fields'SP_TRADED
SP_TRADED

E'Fields'SP_PROJECTED
SP_PROJECTED

fromE'Fields :: E'Fields -> Text

unwrap E'Fields enum

toE'Fields :: Text -> Either String E'Fields

parse E'Fields enum

E'OpTypes

data E'OpTypes

Enum of Text

Constructors

E'OpTypes'Heartbeat
"heartbeat"

E'OpTypes'Authentication
"authentication"

E'OpTypes'MarketSubscription
"marketSubscription"

E'OpTypes'OrderSubscription
"orderSubscription"

fromE'OpTypes :: E'OpTypes -> Text

unwrap E'OpTypes enum

toE'OpTypes :: Text -> Either String E'OpTypes

parse E'OpTypes enum

E'OpTypes2

data E'OpTypes2

Enum of Text

Constructors

E'OpTypes2'Connection
"connection"

E'OpTypes2'Status
"status"

E'OpTypes2'Mcm
"mcm"

E'OpTypes2'Ocm
"ocm"

fromE'OpTypes2 :: E'OpTypes2 -> Text

unwrap E'OpTypes2 enum

toE'OpTypes2 :: Text -> Either String E'OpTypes2

parse E'OpTypes2 enum

E'Ot

data E'Ot

Enum of Text . Order Type - the type of the order (L = LIMIT, MOC = MARKET_ON_CLOSE, LOC = LIMIT_ON_CLOSE)

Constructors

E'Ot'L
L

E'Ot'LOC
LOC

E'Ot'MOC
MOC

fromE'Ot :: E'Ot -> Text

unwrap E'Ot enum

toE'Ot :: Text -> Either String E'Ot

parse E'Ot enum

E'Pt

data E'Pt

Enum of Text . Persistence Type - whether the order will persist at in play or not (L = LAPSE, P = PERSIST, MOC = Market On Close)

Constructors

E'Pt'L
L

E'Pt'P
P

E'Pt'MOC
MOC

fromE'Pt :: E'Pt -> Text

unwrap E'Pt enum

toE'Pt :: Text -> Either String E'Pt

parse E'Pt enum

E'SegmentType

data E'SegmentType

Enum of Text . Segment Type - if the change is split into multiple segments, this denotes the beginning and end of a change, and segments in between. Will be null if data is not segmented

Constructors

E'SegmentType'SEG_START
SEG_START

E'SegmentType'SEG
SEG

E'SegmentType'SEG_END
SEG_END

fromE'SegmentType :: E'SegmentType -> Text

unwrap E'SegmentType enum

toE'SegmentType :: Text -> Either String E'SegmentType

parse E'SegmentType enum

E'Side

data E'Side

Enum of Text . Side - the side of the order. For Line markets a B bet refers to a SELL line and an L bet refers to a BUY line.

Constructors

E'Side'B
B

E'Side'L
L

fromE'Side :: E'Side -> Text

unwrap E'Side enum

toE'Side :: Text -> Either String E'Side

parse E'Side enum

E'Status

data E'Status

Enum of Text

Constructors

E'Status'INACTIVE
INACTIVE

E'Status'OPEN
OPEN

E'Status'SUSPENDED
SUSPENDED

E'Status'CLOSED
CLOSED

fromE'Status :: E'Status -> Text

unwrap E'Status enum

toE'Status :: Text -> Either String E'Status

parse E'Status enum

E'Status2

data E'Status2

Enum of Text

Constructors

E'Status2'ACTIVE
ACTIVE

E'Status2'WINNER
WINNER

E'Status2'LOSER
LOSER

E'Status2'REMOVED
REMOVED

E'Status2'REMOVED_VACANT
REMOVED_VACANT

E'Status2'HIDDEN
HIDDEN

E'Status2'PLACED
PLACED

fromE'Status2 :: E'Status2 -> Text

unwrap E'Status2 enum

toE'Status2 :: Text -> Either String E'Status2

parse E'Status2 enum

E'Status3

data E'Status3

Enum of Text . Status - the status of the order (E = EXECUTABLE, EC = EXECUTION_COMPLETE)

Constructors

E'Status3'E
E

E'Status3'EC
EC

fromE'Status3 :: E'Status3 -> Text

unwrap E'Status3 enum

toE'Status3 :: Text -> Either String E'Status3

parse E'Status3 enum

E'StatusCode

data E'StatusCode

Enum of Text . The status of the last request

Constructors

E'StatusCode'SUCCESS
SUCCESS

E'StatusCode'FAILURE
FAILURE

fromE'StatusCode :: E'StatusCode -> Text

unwrap E'StatusCode enum

toE'StatusCode :: Text -> Either String E'StatusCode

parse E'StatusCode enum

E'Type

data E'Type

Enum of Text

Constructors

E'Type'CLASSIC
CLASSIC

E'Type'FINEST
FINEST

E'Type'LINE_RANGE
LINE_RANGE

fromE'Type :: E'Type -> Text

unwrap E'Type enum

toE'Type :: Text -> Either String E'Type

parse E'Type enum